+++ test/testConfig.js
@@ -1,18 +1,27 @@
 const envs = {};
 
-let rollbar = {
-    accessToken: "someToken",
-    logOnDev: false
-};
-
 envs.development = {
     baseUrl: "http://localhost:3000/api",
-    rollbar
+    uploadCareToken: "someXPTOtoken",
+    rollbarToken: "anotherCrazyToken"
+};
+
+envs.development_firebase = {
+    baseUrl: "https://remotehost-dev.herokuapp.com:3000/api",
+    uploadCareToken: "someXPTOtoken",
+    rollbarToken: "anotherCrazyToken"
+};
+
+envs.staging = {
+    baseUrl: "https://remotehost-staging.herokuapp.com:3000/api",
+    uploadCareToken: "someXPTOtoken",
+    rollbarToken: "anotherCrazyToken"
 };
 
 envs.production = {
-    baseUrl: "http://remote:3000/api",
-    rollbar
+    baseUrl: "https://remotehost-prod.herokuapp.com:3000/api",
+    uploadCareToken: "someXPTOtoken",
+    rollbarToken: "anotherCrazyToken"
 };
 
 envs.staging = envs.production;